+++ packages/oauth/oauth-provider-ui/src/data/password.ts
@@ -1,4 +1,4 @@
-import { useLingui } from '@lingui/react/macro'
+import { msg } from '@lingui/core/macro'
 import { useMutation } from '@tanstack/react-query'
 import {
   ConfirmResetPasswordInput,
@@ -10,8 +10,7 @@ import { WithOptionalLocale } from '#/lib/api.ts'
 
 export function useResetPasswordRequest() {
   const api = useApi()
-  const { t } = useLingui()
-  const { notify } = useNotificationsContext()
+  const { notify, notifyError } = useNotificationsContext()
 
   return useMutation({
     async mutationFn(data: WithOptionalLocale<InitiatePasswordResetInput>) {
@@ -19,17 +18,14 @@ export function useResetPasswordRequest() {
     },
     onSuccess(_data, _variables, _context) {
       notify({
-        variant: 'success',
-        title: t`Password reset email sent`,
-        description: t`Check your inbox.`,
+        title: msg`Password reset email sent`,
+        description: msg`Check your inbox.`,
       })
     },
     onError(error, _variables, _context) {
-      console.error('Failed to request password reset', error)
-      notify({
-        variant: 'error',
-        title: t`Failed to request password reset`,
-        description: t`Please check the email address and try again.`,
+      notifyError(error, {
+        title: msg`Failed to request password reset`,
+        description: msg`Please check the email address and try again.`,
       })
     },
   })
@@ -37,8 +33,7 @@ export function useResetPasswordRequest() {
 
 export function useResetPasswordConfirm() {
   const api = useApi()
-  const { t } = useLingui()
-  const { notify } = useNotificationsContext()
+  const { notify, notifyError } = useNotificationsContext()
 
   return useMutation({
     async mutationFn(data: ConfirmResetPasswordInput) {
@@ -46,17 +41,14 @@ export function useResetPasswordConfirm() {
     },
     onSuccess(_data, _variables, _context) {
       notify({
-        variant: 'success',
-        title: t`Password reset successful`,
-        description: t`You can now sign in with your new password.`,
+        title: msg`Password reset successful`,
+        description: msg`You can now sign in with your new password.`,
       })
     },
     onError(error, _variables, _context) {
-      console.error('Failed to reset password', error)
-      notify({
-        variant: 'error',
-        title: t`Failed to reset password`,
-        description: t`Please check your reset code and try again.`,
+      notifyError(error, {
+        title: msg`Failed to reset password`,
+        description: msg`Please check your reset code and try again.`,
       })
     },
   })